Seguimos movimiendo los archivos
Cree una carpeta directamente pensada para los tipos de datos basicos que voy creando y una para el quadtree
La idea de esto es que despues cuando mas cosas se vayan creando, si necesito estos tipos de datos esten en un lugar que tenga sentido
Tambien replique la misma estructura de carpetas en los test para que tenga sentido donde encontrar cada cosa

+++ src/herramientas/quadtree/node.cpp
@@ -1,5 +1,5 @@
 
-#include "headers/node.h"
+#include "node.h"
 
 Node::Node(const Rectangulo &nueva_area)
     : c_area(nueva_area), c_cant_entidades(0), c_dividido(false)
@@ -31,8 +31,8 @@ int Node::calcular_index(const Entidad &entidad) const
 {
     Vector2 delta = entidad.c_pos - c_area.c_pos;
 
-    int index_x = (delta.x == 0) ? 0 : static_cast<int>((1 + delta.x / abs(delta.x)) / 2);
-    int index_y = (delta.y == 0) ? 0 : static_cast<int>((1 - delta.y / abs(delta.y)) / 2);
+    int index_x = (delta.x == 0) ? 0 : static_cast<int>((1 + delta.x / abs((int)delta.x)) / 2);
+    int index_y = (delta.y == 0) ? 0 : static_cast<int>((1 - delta.y / abs((int)delta.y)) / 2);
 
     return (index_x + 2 * index_y);
 }
